Replace item.add_report_section() with long_repr.addsection()
Commit cfedc4 does the reverse change and introduces the follwing bug.
When used together with pytest junitxml plugin, the XML report does not
contain the captured log output, since junit-xml does not know about
the 'log' section.

Also, with this change when a test fails we will see the logs only from the
method that caused the failure, which means if a test failed in teardown
phase we will not see logs from setup and call phases.